third_party/blink/web_tests/platform/generic/external/wpt/scroll-animations/css/at-scroll-timeline-responsiveness-from-endpoint-expected.txt

This is a testharness.js-based test.
FAIL Test that the scroll animation is still responsive after moving from 100% assert_equals: expected "100px" but got "50px"
Harness: the test ran to completion.
